src/server/pyproject.toml
[project]
name = "federated-collection-discovery"
dynamic = ["version"]
description = "An application for discovering collections from multiple STAC APIs and/or CMRs"
authors = [
    {name = "Henry Rodman", email = "henry@developmentseed.org"},
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pystac-client>=0.8.4",
    "stac-fastapi-types>=3.0.0",
    "pydantic-settings>=2.3.4",
    "black>=24.4.2",
    "httpx>=0.27.2",
    "python-cmr>=0.11.0",
    "fastapi>=0.111.1",
]

[project.optional-dependencies]
aws_lambda = ["mangum>=0.17.0"]
uvicorn = ["uvicorn>=0.30.1"]
test = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "requests-mock>=1.12.1",
    "asgi-lifespan>=2.1.0",
    "pytest-recording>=0.13.2",
    "pytest-cov>=5.0.0",
]
dev = [
    "jupyterlab>=4.2.5",
    "pre-commit>=4.0.1",
    "ruff>=0.6.9",
    "types-requests>=2.32.0.20241016",
]

[project.urls]
homepage = "https://github.com/developmentseed/federated-collection-discovery"

[tool.setuptools.dynamic]
version = { attr = "federated_collection_discovery.__version__" }

[tool.pytest.ini_options]
# filterwarnings = "ignore:Server does not conform to COLLECTION_SEARCH:UserWarning"
markers = "vcr: records network activity"
addopts = "-Werror --block-network --cov=federated_collection_discovery --cov-report=term-missing --cov-report=xml"

[tool.setuptools]
packages = ["federated_collection_discovery"]

[tool.isort]
profile = "black"

[tool.mypy]
explicit_package_bases = true
namespace_packages = true
show_error_codes = true
strict = true
mypy_path = "src"

[tool.ruff]
lint.select = ["E", "F", "I"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
